cloud_functions/input_pubsub/requirements.txt
google-cloud-datastore
google-cloud-storage
google-cloud-pubsub
google-api-python-client
opencv-python
numpy
requests
Pillow